requirements.txt
streamlit
pdfplumber
PyPDF2
pillow
pytesseract
python-dotenv>=1.0.0
